fix(react): richtext table and custom attribute casing (#292)
I’ve decided to add as many attributes as possible to the map of HTML
attributes that must be converted to camelCase for React. Although only
a couple of attributes are supported by default, it is possible to allow
custom attributes on links, so in practice, there could be all the
double-word HTML attributes. I’ve also added colwidth, although this is
not a standard attribute, because the table module seems to set it.

Fixes WDX-82
Fixes #239