fix(cli): removing fields via migrations
Since `Object.assign()` is only additive, we have introduced a check to
support subtractive migrations: if a property is absent from the
migrated content, it will now be removed from the target object as well.

`modified` was renamed to `processed` so it better describes the fact
that although we handled a migration, it might not have affected the
content.