fix(cli): add support for `--filename` flag to allow custom file name on generated component types (#326)
#### Description
This PR enables the `--filename` flag to be used for the components type
files.
Previously, the option existed but the CLI command did not handle it.

#### What’s changed
- Added support for `--filename` flag in the types generation flow.
- Introduced unit tests to ensure correct behavior
- Default filename will be `storyblok-components.d.ts`, custom will
override it like so: `custom-name.d.ts`
- Update docs

Notes
- This gives users more control over the naming for the generated
components type files.
- A potential improvement could be to allow the consumer have custom
name for the `storyblok.d.ts` file as well. However that file is
Storyblok specific types and is not as important to have with a custom
name as of now

---------

Co-authored-by: Niklas Rydkvist <niklasrydkvist@Niklass-MBP.lan>
Co-authored-by: Niklas Rydkvist <niklas.rydkvist@hm.com>