fix(cli): retry mechanism when multiple systems query the MAPI (#358)
When multiple systems hit the MAPI with the same access token, the retry
mechanisms quickly gave up. By decreasing the semaphore batch size and
increasing the max retries, we make such situations less likely.
However, it is still possible to run into limits when many systems query
the MAPI with the same access token in parallel.

See the following article for the rational behind the full jitter retry
delay:
https://aws.amazon.com/blogs/architecture/exponential-backoff-and-jitter/