fix(react): update nextjs peer dep and re-structure playgrounds (#361)
After investigating, testing and playing with it it, seems adding the
peer dep was all needed.

At the same time, I took the chance to rename the playgrounds and fix a
few minor things:
- `next13-app-router` -> `next-e2e-testing` (it was the one used for it,
so now it's clearer that it's the one used for testing, and should be
changed only when automated tests change, not for manual testing - just
as we have it in the Nuxt SDK)
- `next15` -> `next-latest` (so we don't need to rename it in every
major release. Also, upgraded to Next 16)
- `next15-static` -> `next-latest-static` (same)


Closes #360