fix(richtext)!: HTML and Markdown parsing for links and edge cases
Parsing did not work correctly for certain combinations of `<a>` and
nested tags inside or around it.

To ensure parsing is aligned with the Tiptap editor format, we now use
the Tiptap editor's `generateJSON` command, which gives us
HTML-to-richtext conversion out of the box.

BREAKING CHANGE: Instead of our own resolver format, we now allow users
to override the Tiptap extensions used for parsing.

Fixes WDX-141